refactor: remove RegisterFormStep2 component and adjust registration flow to streamline steps
