+++ components/hackathons/registration-form/RegisterFormStep3.tsx
@@ -7,10 +7,21 @@ import {
   FormLabel,
   FormMessage,
 } from "@/components/ui/form";
-
+import {
+  Select,
+  SelectContent,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from "@/components/ui/select";
 import { useFormContext } from "react-hook-form";
 import { RegisterFormValues } from "./RegistrationForm"; 
 import { Checkbox } from "@/components/ui/checkbox";
+
+const hackathonParticipationOptions = [
+  { value: "yes", label: "Yes" },
+  { value: "no", label: "No" },
+];
 
 interface RegisterFormStep3Props {
   isOnlineHackathon: boolean;
@@ -21,15 +32,44 @@ export function RegisterFormStep3({ isOnlineHackathon }: RegisterFormStep3Props)
 
   return (
     <>
-      {/* Step 3: Terms & Agreements */}
+      {/* Step 2: Terms & Agreements */}
    
       <div className="mb-6">
-        <h3 className="text-lg font-semibold text-foreground">Step 3: Terms & Agreements</h3>
+        <h3 className="text-lg font-semibold text-foreground">Step 2: Terms & Agreements</h3>
         <p className="text-zinc-400">Review and agree to the terms to complete your registration. For information about our privacy practices and commitment to protecting your privacy, please review our <a href="https://www.avax.network/privacy-policy" target="_blank" rel="noopener noreferrer" className="text-blue-500 underline"> Avalanche Privacy Policy. </a></p>
         <div className="w-full h-px bg-zinc-300 mt-2" /> 
       </div>
       <div className="space-y-6">
+        <FormField
+          control={form.control}
+          name="hackathon_participation"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>
+                Have you participated in any other hackathons before? (Optional)
+              </FormLabel>
+              <Select onValueChange={field.onChange} value={field.value}>
+                <FormControl>
+                  <SelectTrigger className="text-zinc-600">
+                    <SelectValue placeholder="Select an option (optional)" />
+                  </SelectTrigger>
+                </FormControl>
+                <SelectContent className="bg-white dark:bg-black border-gray-300 dark:border-zinc-600 text-black dark:text-zinc-600 rounded-md shadow-md">
+                  {hackathonParticipationOptions.map((option) => (
+                    <SelectItem key={option.value} value={option.value}>
+                      {option.label}
+                    </SelectItem>
+                  ))}
+                </SelectContent>
+              </Select>
+              <FormMessage className="text-zinc-600">
+                Let us know if this is your first hackathon or if you have prior experience.
+              </FormMessage>
+            </FormItem>
+          )}
+        />
 
+        <div className="w-full h-px bg-zinc-300" />
 
         <FormField
           control={form.control}
